Updates system table reckognition for Firebird Extension to only identify RDB$ and MON$ tables, rather than all tables with $ in the name.
MON$ tables are not described as system tables, but rather as monitoring tables. It makes sense to include them as system tables regardless.

Firebird reference manual states:

> System table identifiers all begin with the prefix RDB$

and
> [...] monitoring tables. The definitions of these tables are always present in the database, all named with the prefix MON$

Normal table names are allowed to contain $ too. Table names are restricted only by uniqueness and length:

> Name (identifier) for the table. It may consist of up to 31 characters and must be unique in the database.

The absolute best practice would be to lookup the  of  for the table in question.

Links to the reference quotes:

- https://www.firebirdsql.org/file/documentation/reference_manuals/fblangref25-en/html/fblangref25-appx04-systables.html
- https://www.firebirdsql.org/file/documentation/reference_manuals/fblangref25-en/html/fblangref25-appx05-montables.html
- https://www.firebirdsql.org/file/documentation/reference_manuals/fblangref25-en/html/fblangref25-ddl-tbl.html


Former-commit-id: 16a59453ba512d884cf93a5a48badcfc9fa8e2fe