#6536 #10184 macOS Horizontal Scroll Issue
The method scrollHorizontally() is called on on event SWT.MouseHorizontalWheel. In my testing, it seems like this event is being raised on macOS, but not Windows. The method overrides normal horizontal scrolling and forces either the leftmost or rightmost visible column to be fully visible - depending on whether the user is scrolling to the left or right. This causes scrolling to feel "jumpy". If the user has columnScrolling disabled (i.e. Smooth Scrolling enabled), this method probably shouldn't be called.

The value for HORZ_SCROLL_INCREMENT is only used if columnScrolling is disabled (Smooth Scrolling enabled) and only impacts the UI when responding to a horizontal mouse wheel. I tested a range of values on a 2019 MacBook Pro and a Microsoft Surface Laptop 4, and a value of 12 gave the closest approximation to scrolling in applications like Excel. Behavior was unchanged when using a basic wired mouse.